Add the ability to print examples
You can now specify PRINT_EXAMPLES=true as an environment variable and it will
print all examples provided by the `shouldShowExample` function in the code.
Those examples all go to a file called soak-examples.txt (which gets clobbered
if it's already there).

Maybe in the future you'll be able to specify a predicate for which examples to
print, but for now it's just in the code.

Table printing code pretty much all taken from decaffeinate error reporting.